Error in allocating memory for conditional fields
Fixed an issue where memory allocation for arguments with conditional fields may be incorrect.
For example, next ptr:

structure_test_cond{
	cond int8[1:2]
	ut union_test
} [packed]

union_test [
	first array[int64,8] (if[value[structure_test_cond:cond] == 1])
	default int8
] [varlen]

Now, when generating values, the possible values are generated first,
and then the function for generating conditional fields is executed.
(patchConditionalFields).
In this case, the conditional fields are filled with the default value.
In the example, this value is default, which results in the generation and allocation of memory for a 2-byte structure.
After the execution of patchConditionalFields, the structure may change its size to 41 bytes,
but the memory will be allocated during the generation of the pointer (for a 2-byte structure),
resulting in the next pointer pointing to already used memory.

To fix this error, it is suggested to move the allocation of memory for
the pointers after executing patchConditionalFields.